Add non-null userId check
Bad request when userId is null/empty/whitespace to task endpoint

+++ API/API/Controllers/TaskController.cs
@@ -18,8 +18,14 @@ public class TaskController : ControllerBase
     }
 
     [HttpGet]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<ActionResult<List<TodoTaskDto>>> Get([FromHeader] string userId)
     {
+        if (string.IsNullOrWhiteSpace(userId))
+        {
+            return BadRequest();
+        }
         var tasks = (await _todoService.GetUserTasksAsync(userId)).Select(t => new TodoTaskDto
         {
             Id = t.Id,
@@ -34,8 +40,13 @@ public class TaskController : ControllerBase
     [HttpPost]
     [ProducesResponseType(StatusCodes.Status201Created)]
     [ProducesResponseType(StatusCodes.Status409Conflict)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<ActionResult<TodoTaskDto>> Create([FromHeader] string userId, [FromBody] CreateTodoTaskDto dto)
     {
+        if (string.IsNullOrWhiteSpace(userId))
+        {
+            return BadRequest();
+        }
         try
         {
             var createTask = await _todoService.CreateTaskAsync(dto, userId);
@@ -58,8 +69,10 @@ public class TaskController : ControllerBase
     [HttpGet("{id}")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<ActionResult<TodoTaskDto>> GetTodoTask([FromHeader] string userId, string id)
     {
+        if (string.IsNullOrWhiteSpace(userId)) { return BadRequest(); }
         var task = await _todoService.GetTaskAsync(userId, id);
         if (task == null)
         {
@@ -78,8 +91,10 @@ public class TaskController : ControllerBase
     [HttpDelete("{id}")]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<ActionResult> DeleteTodoTask([FromHeader] string userId, string id)
     {
+        if (string.IsNullOrWhiteSpace(userId)) { return BadRequest(); }
         try
         {
             await _todoService.DeleteTaskAsync(id, userId);
@@ -94,8 +109,10 @@ public class TaskController : ControllerBase
     [HttpPatch("{id}/completed")]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<ActionResult> MarkTodoTaskCompleted([FromHeader] string userId, string id)
     {
+        if (string.IsNullOrWhiteSpace(userId)) { return BadRequest(); }
         try
         {
             await _todoService.UpdateTaskAsCompletedAsync(id, userId);
